requirements.txt
ipdb
opencv-python
pip-chill
pytube3
scikit-image
unidecode
